change para seq

+++ airbase-5g.sh
@@ -1,21 +1,30 @@
 #!/bin/bash
 maclist=$1
-essid=0
-location=$(basename  $1  |grep -oE ^.*[\.])
+#essid=$(cat /dev/urandom | head -n 10 | md5sum | head -c 19)
+location=$(basename  $1  |grep -oE ^.*[\.]?)
 spoof(){
+	airmon-ng check kill
 	for mac in $(grep -E ^[^#] $maclist)
 	do
 		let i=i+1	
-		let essid=essid+111 
+#		let essid=essid+111 
+		essid=$(cat /dev/urandom | head -n 10 | md5sum | head -c 11)
 		sleep 1
 		clear
 		airbase-ng -e $essid -c 36 -a "$mac" wlan0 & 
-		echo "Starting $location$i:  $essid ..."
+		echo "Starting $location-$i "
+		echo "SSID:-- $essid "
+	       	echo "MAC:--- $mac"
 	done
 }
+if [ $UID -ne 0 ];then
+	echo "This Script Must Run As Root .. "
+	exit
+fi
 
 pkill airbase-ng
 ifconfig  -a |grep wlan0 
+
 if [ $? != 0 ];then
 	echo "Device not found"
 	exit
